auth/auth_presentation/src/main/java/com/acidtango/auth_presentation/viewModels/RegistrationVM.kt
package com.acidtango.auth_presentation.viewModels

import androidx.compose.runtime.getValue
import androidx.compose.runtime.mutableStateOf
import androidx.compose.runtime.setValue
import androidx.lifecycle.ViewModel
import dagger.hilt.android.lifecycle.HiltViewModel
import javax.inject.Inject

@HiltViewModel
class RegistrationVM
@Inject constructor() : ViewModel() {

    var nickName by mutableStateOf("")
        private set

    var email by mutableStateOf("")
        private set

    var password by mutableStateOf("")
        private set

    fun onNickNameChange(nickName: String) {
        this.nickName = nickName
    }

    fun onEmailChange(email: String) {
        this.email = email
    }

    fun onPasswordChange(password: String) {
        this.password = password
    }
}
